refactor: split Postgres store by repository slice
